refactor: relocate serde functions
