+++ build.rs
@@ -25,11 +25,11 @@ fn main() {
         .type_attribute(".", "#[derive(serde::Serialize,serde::Deserialize)]")
         .field_attribute(
             "sxt.core.ProverResponse.exec_time",
-            "#[serde(default, serialize_with = \"crate::base::duration_serde::serialize\", deserialize_with = \"crate::base::duration_serde::deserialize\")]",
+            "#[serde(default, serialize_with = \"crate::base::serde::duration_serde::serialize\", deserialize_with = \"crate::base::serde::duration_serde::deserialize\")]",
         )
         .field_attribute(
             "sxt.core.ProofAndResultForCommitments.exec_time",
-            "#[serde(default, serialize_with = \"crate::base::duration_serde::serialize\", deserialize_with = \"crate::base::duration_serde::deserialize\")]",
+            "#[serde(default, serialize_with = \"crate::base::serde::duration_serde::serialize\", deserialize_with = \"crate::base::serde::duration_serde::deserialize\")]",
         )
         .type_attribute(".", "#[allow(clippy::all)]")
         .build_client(cfg!(feature = "prover-client"))
