+++ src/base/serde/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod duration_serde;
+pub(crate) mod hex;
